Rewrite test (+ new test) of animate_depth_timeseries
New test dependency on pillow

Move reference images to test-data/baseline

Rendering of png files from plotly animations can vary slightly if the library versions
are not identical. So instead of comparing bit for bit, we read the images and calculate
the root mean squared difference. If this is less than a threshold the images match.